Create NumbersFragment
