+++ app/src/main/java/com/example/mf4z/fulfulde/NumbersFragment.java
@@ -1,0 +1,173 @@
+package com.example.mf4z.fulfulde;
+
+
+import android.content.Context;
+import android.media.AudioManager;
+import android.media.MediaPlayer;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ListView;
+import android.widget.TextView;
+
+import java.util.ArrayList;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * create an instance of this fragment.
+ */
+public class NumbersFragment extends Fragment {
+    //Declaring Media Player variable
+    private MediaPlayer mMediaPlayer;
+
+    //Delaring Audio Manager for use in the entire class
+    private AudioManager mAudioManager;
+
+    private MediaPlayer.OnCompletionListener mCompletionListener = new MediaPlayer.OnCompletionListener() {
+        @Override
+        public void onCompletion(MediaPlayer mediaPlayer) {
+            // Now that the sound file has finished playing, release the media player resources.
+            releaseMediaPlayer();
+        }
+    };
+
+
+    private AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
+        @Override
+        public void onAudioFocusChange(int focusChange) {
+            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT ||
+                    focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK) {
+                // The AUDIOFOCUS_LOSS_TRANSIENT case means that we've lost audio focus for a
+                // short amount of time. The AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK case means that
+                // our app is allowed to continue playing sound but at a lower volume. We'll treat
+                // both cases the same way because our app is playing short sound files.
+
+                // Pause playback and reset player to the start of the file. That way, we can
+                // play the word from the beginning when we resume playback.
+                mMediaPlayer.pause();
+                mMediaPlayer.seekTo(0);
+            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
+                // The AUDIOFOCUS_GAIN case means we have regained focus and can resume playback.
+                mMediaPlayer.start();
+            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
+                // The AUDIOFOCUS_LOSS case means we've lost audio focus and
+                // Stop playback and clean up resources
+                releaseMediaPlayer();
+            }
+        }
+    };
+
+
+    public NumbersFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        View rootView = inflater.inflate(R.layout.word_list, container, false);
+
+        /** TODO: Insert all the code from the NumberActivity’s onCreate() method after the setContentView method call */
+
+        //Create and Setup
+        mAudioManager = (AudioManager) getActivity().getSystemService(Context.AUDIO_SERVICE);
+
+        //Creating a list of words with an ArrayList of object Word
+        final ArrayList<Word> words = new ArrayList<>();
+
+        //using nameOfList.add() to add objects to the list
+        words.add(new Word("Go'o","One",R.drawable.number_one,R.raw.num1));
+        words.add(new Word("Didi","Two",R.drawable.number_two,R.raw.num2));
+        words.add(new Word("Tati","Three",R.drawable.number_three,R.raw.num3));
+        words.add(new Word("Nai","Four",R.drawable.number_four,R.raw.num4));
+        words.add(new Word("Jui","Five",R.drawable.number_five,R.raw.num5));
+        words.add(new Word("Jego","Six",R.drawable.number_six,R.raw.num6));
+        words.add(new Word("Jedidi","Seven",R.drawable.number_seven,R.raw.num7));
+        words.add(new Word("Jetati","Eight",R.drawable.number_eight,R.raw.num8));
+        words.add(new Word("Jenai","Nine",R.drawable.number_nine,R.raw.num9));
+        words.add(new Word("Sappo","Ten",R.drawable.number_ten,R.raw.num10));
+
+
+        //Instantiating a new object of WordAdapter adepter,takes in thr context,object list,theme color
+        WordAdapter adapter = new WordAdapter(getActivity(),words,R.color.colorGreen);
+
+        //Declaring and initialising the list view
+        ListView listView = rootView.findViewById(R.id.list);
+
+
+        //Setting the adapter to the listview
+        listView.setAdapter(adapter);
+
+
+        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+
+                // Release the media player if it currently exists because we are about to
+                // play a different sound file
+                releaseMediaPlayer();
+
+                // Get the {@link Word} object at the given position the user clicked on
+                Word word = words.get(position);
+
+                //Log.v("NumbersActivity", "Current word: " + word);
+
+// Request audio focus for playback
+                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
+                        // Use the music stream.
+                        AudioManager.STREAM_MUSIC,
+                        // Request permanent focus.
+                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
+
+                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+
+                    // We have audio focus now
+
+                    //Creating a media player with a song
+                    mMediaPlayer = MediaPlayer.create(getActivity(),word.getMwordSound());
+                    //play sound
+                    mMediaPlayer.start();
+
+                    // Setup a listener on the media player, so that we can stop and release the
+                    // media player once the sound has finished playing.
+                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
+
+                }
+
+            }
+        });
+
+        return rootView;
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        // When the activity is stopped, release the media player resources because we won't
+        // be playing any more sounds.
+        releaseMediaPlayer();
+    }
+
+    private void releaseMediaPlayer() {
+        // If the media player is not null, then it may be currently playing a sound.
+        if (mMediaPlayer != null) {
+            // Regardless of the current state of the media player, release its resources
+            // because we no longer need it.
+            mMediaPlayer.release();
+
+            // Set the media player back to null. For our code, we've decided that
+            // setting the media player to null is an easy way to tell that the media player
+            // is not configured to play an audio file at the moment.
+            mMediaPlayer = null;
+
+            // Regardless of whether or not we were granted audio focus, abandon it. This also
+            // unregisters the AudioFocusChangeListener so we don't get anymore callbacks.
+            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
+        }
+    }
+}
