+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,7 @@
     <string name="familyM">Family Members</string>
     <string name="colors">Colors</string>
     <string name="phrases">Phrases</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
